fix: merge intake_by_uuid and modules_by_uuid plugins into a single on_page_read_source hook (as mkdocs forbids multiple on_page_read_source hooks)
